добавлены контролеры, севрисы, и дто

+++ src/main/java/com/wanted/needswork/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.wanted.needswork.repository;
 import com.wanted.needswork.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository <User,Integer> {
+import java.math.BigInteger;
+
+public interface UserRepository extends JpaRepository <User, BigInteger> {
 
 }
